use 3 vertices per data vertex
This works fine for making a "knee" to enforce the miter limit.
But, where the miter limit does not need to be enforced (doesn't
need to be kneed ;-), we don't need that extra vertex, and I can't
figure out where to put it which will not result in some trouble
caused by the creation of a degenerate triangle.