replace timeScale property with timeSpan
timeScale would often need to be calculated based on width anyway.
And soon, the model will need to worry about keeping a certain time period
instead of a certain number of samples: the periodicity is allowed to be
irregular whereas the graph always shows a certain span of time.
So, the model needs to know the time span too.  We might as well be
consistent.