influx: fix the timespan
LineGraphModel::appendVertices() tries to enforce it for clipping
purposes; so we have to start by setting timeSpan to INT_MAX to
prevent that enforcement.  Only after iterating through all the
JSON values do we know the current time span, because
InfluxQuery.timeConstraint is a string for influx to parse, not
for us to parse ahead of time.  Could also be fixed by adding
a timeSpan property and requiring that InfluxQuery can only get
the most recent timeSpan seconds of data.  But that's less flexible
for the future.  Maybe rather have timeSpan and endTime?