LmSensors min/max subfeatures can be undefined
That is, sensors_get_subfeature(chip, feature, SENSORS_SUBFEATURE_IN_MIN/MAX)
are returning null for cpu0_vid voltage, whatever that is, so have to check
before accessing the subfeature.  The "sensors" client is also not showing
min/max values for that voltage.  (And its value seems to always be 0
anyway so it's not even an interesting one.)